muscle ellipsoid 2d forward: bool mask operation

+++ virtual_nematode/networks/snn/forward.py
@@ -128,7 +128,7 @@ class SNNCell(torch.nn.Module):
         activation: cell activation, (batch_size, cell_count)
         proprioception: (batch_size, proprioception_size)
         """
-        w_c = self.w_c.abs() * self.ex_mask_c - self.w_c.abs() * self.in_mask_c + self.w_c * (self.mask_c - self.ex_mask_c - self.in_mask_c)
+        w_c = self.w_c.abs() * self.ex_mask_c - self.w_c.abs() * self.in_mask_c + self.w_c * (self.mask_c.float() - self.ex_mask_c.float() - self.in_mask_c.float())
         synapse_input = torch.mm(activation, w_c)  # chemical synapse input, (batch_size, cell_count)
         delta_state = state.unsqueeze(dim=2).repeat(1, 1, self.n) - state.unsqueeze(dim=1).repeat(1, self.n, 1)
         w_g = self.w_g.abs() * self.mask_g
